requirements.txt
streamlit
pdfplumber
chromadb
cohere
python-dotenv
